requirements.txt
alembic==1.7.5
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
defusedxml==0.7.1
dnspython==2.2.0
dominate==2.6.0
email-validator==1.1.3
Flask==2.0.2
Flask-Admin==1.6.0
Flask-DotEnv==0.1.2
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
odfpy==1.4.1
pycparser==2.21
PyMySQL==1.0.2
pyOpenSSL==22.0.0
python-dotenv==0.19.2
requests==2.27.1
SQLAlchemy==1.4.31
urllib3==1.26.8
visitor==0.1.3
waitress==2.0.0
Werkzeug==2.0.2
WTForms==3.0.1
